1-listint_len.c

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,13 +1,13 @@
 #include "lists.h"
 
-/*
+/**
  * listint_len - find length of linked list
  * @h: linked list
  * Return: length of linked list
  */
 size_t listint_len(const listint_t *h)
 {
-	int length = 0;
+	size_t length = 0;
 
 	while (h != NULL)
 	{
